Move classes in to sub-packages of the main package to allow default component scanning of sub-packages
