+++ src/main/java/com/johnny/spring/springcoredemo/SpringcoredemoApplication.java
@@ -3,6 +3,15 @@ package com.johnny.spring.springcoredemo;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+/**
+ * @SpringBootApplication is composed of the following annotations:
+ * @EnableAutoConfiguration - Enables Spring Boot's auto-configuration
+ * suppoort
+ * @ComponentScan - Enables component scanning of current package.
+ * Also recursively scans sub-packages
+ * @Configuration - Able to regist extra beans with @Bean or import
+ * other configuration classes
+ */
 @SpringBootApplication
 public class SpringcoredemoApplication {
