+++ src/main/java/com/johnny/spring/springcoredemo/common/Coach.java
@@ -1,4 +1,4 @@
-package com.johnny.spring.springcoredemo;
+package com.johnny.spring.springcoredemo.common;
 
 
 public interface Coach {
