+++ src/main/java/com/johnny/spring/springcoredemo/common/CricketCoach.java
@@ -1,4 +1,4 @@
-package com.johnny.spring.springcoredemo;
+package com.johnny.spring.springcoredemo.common;
 
 import org.springframework.stereotype.Component;
