+++ src/main/java/com/johnny/spring/springcoredemo/rest/DemoController.java
@@ -1,5 +1,6 @@
-package com.johnny.spring.springcoredemo;
+package com.johnny.spring.springcoredemo.rest;
 
+import com.johnny.spring.springcoredemo.common.Coach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
